Extend PublicActivity with CommonWithSystem
this adds #create_system_activity which generates activities where the owner_type is 'System' and the owner_id is -1. Inspired by the community bot @stackoverflow who is also -1.

Tests are in place and the normal functionality of PublicActivity is assured, still it's monkey patched a bit, so update public_activity with caution